fix(platform-browser): `DomEventsPlugin` should always be the last plugin to be called for `supports()`.
This fixes the issues when `BrowserModule` is not the first module imported.

Fixes #37149 #37850
