+++ packages/platform-browser/src/dom/events/event_manager.ts
@@ -17,6 +17,13 @@ import {
 
 import {RuntimeErrorCode} from '../../errors';
 
+import type {EventManagerPlugin} from './event_manager_plugin';
+
+// Temporart re-export until we can change files that import from this file in G3.
+export {EventManagerPlugin} from './event_manager_plugin';
+
+import {DomEventsPlugin} from './dom_events';
+
 /**
  * The injection token for plugins of the `EventManager` service.
  *
@@ -47,7 +54,15 @@ export class EventManager {
     plugins.forEach((plugin) => {
       plugin.manager = this;
     });
-    this._plugins = plugins.slice().reverse();
+
+    const otherPlugins = plugins.filter((p) => !(p instanceof DomEventsPlugin));
+    this._plugins = otherPlugins.slice().reverse();
+
+    // DomEventsPlugin.supports() always returns true, it should always be the last plugin.
+    const domEventPlugin = plugins.find((p) => p instanceof DomEventsPlugin);
+    if (domEventPlugin) {
+      this._plugins.push(domEventPlugin);
+    }
   }
 
   /**
@@ -97,35 +112,4 @@ export class EventManager {
     this._eventNameToPlugin.set(eventName, plugin);
     return plugin;
   }
-}
-
-/**
- * The plugin definition for the `EventManager` class
- *
- * It can be used as a base class to create custom manager plugins, i.e. you can create your own
- * class that extends the `EventManagerPlugin` one.
- *
- * @publicApi
- */
-export abstract class EventManagerPlugin {
-  // TODO: remove (has some usage in G3)
-  constructor(private _doc: any) {}
-
-  // Using non-null assertion because it's set by EventManager's constructor
-  manager!: EventManager;
-
-  /**
-   * Should return `true` for every event name that should be supported by this plugin
-   */
-  abstract supports(eventName: string): boolean;
-
-  /**
-   * Implement the behaviour for the supported events
-   */
-  abstract addEventListener(
-    element: HTMLElement,
-    eventName: string,
-    handler: Function,
-    options?: ListenerOptions,
-  ): Function;
 }
