+++ packages/platform-browser/src/platform-browser.ts
@@ -19,7 +19,8 @@ export {Title} from './browser/title';
 export {disableDebugTools, enableDebugTools} from './browser/tools/tools';
 export {By} from './dom/debug/by';
 export {REMOVE_STYLES_ON_COMPONENT_DESTROY} from './dom/dom_renderer';
-export {EVENT_MANAGER_PLUGINS, EventManager, EventManagerPlugin} from './dom/events/event_manager';
+export {EVENT_MANAGER_PLUGINS, EventManager} from './dom/events/event_manager';
+export {EventManagerPlugin} from './dom/events/event_manager_plugin';
 export {
   HAMMER_GESTURE_CONFIG,
   HAMMER_LOADER,
