+++ packages/platform-browser/test/dom/events/event_manager_spec.ts
@@ -9,7 +9,8 @@
 import {ɵgetDOM as getDOM} from '@angular/common';
 import {NgZone} from '@angular/core';
 import {DomEventsPlugin} from '../../../src/dom/events/dom_events';
-import {EventManager, EventManagerPlugin} from '../../../src/dom/events/event_manager';
+import {EventManager} from '../../../src/dom/events/event_manager';
+import {EventManagerPlugin} from '../../../src/dom/events/event_manager_plugin';
 
 import {TestBed} from '@angular/core/testing';
 import {isNode, createMouseEvent, el} from '@angular/private/testing';
